api/requirements.txt
Django==3.0.8
djangorestframework==3.11.0
django-allauth==0.44.0
django-cors-headers==3.4.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
coreapi==2.3.3
PyMySQL==1.0.2
django-filter==2.4.0
Pillow==7.2.0
django-cleanup==5.1.0
django-environ==0.4.5